+++ draw-service/src/main/java/com/fcfsdraw/draw/entry/domain/DrawFailReason.java
@@ -2,5 +2,6 @@ package com.fcfsdraw.draw.entry.domain;
 
 public enum DrawFailReason {
     SOLD_OUT,
-    ALREADY_ENTERED
+    ALREADY_ENTERED,
+    PAYMENT_FAILED
 }