+++ draw-service/src/main/java/com/fcfsdraw/draw/payment/client/PaymentFailedException.java
@@ -1,0 +1,8 @@
+package com.fcfsdraw.draw.payment.client;
+
+public class PaymentFailedException extends RuntimeException {
+
+    public PaymentFailedException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}